+++ Changelog.md
@@ -11,6 +11,7 @@
     -Now using Service Worker instead of Background Page
     -Added a Timeline for current Anime Releases
     -Button to visually remove Not Yet Aired from Plan to Watch List
+    -Added Global Storage
 
 6.5.7: 26.10.2021
     -Fixed some Players not skipping on 9anime