+++ Changelog.md
@@ -12,6 +12,7 @@
     -Added a Timeline for current Anime Releases
     -Button to visually remove Not Yet Aired from Plan to Watch List
     -Added Global Storage
+    -Changed 9anime Pattern
 
 6.5.7: 26.10.2021
     -Fixed some Players not skipping on 9anime