+++ Changelog.md
@@ -13,6 +13,8 @@
     -Button to visually remove Not Yet Aired from Plan to Watch List
     -Added Global Storage
     -Changed 9anime Pattern
+    -Added License
+    -Fixed Bugs
 
 6.5.7: 26.10.2021
     -Fixed some Players not skipping on 9anime